+++ pbkdf2/tests/lib.rs
@@ -1,8 +1,6 @@
 use pbkdf2;
 
 use hmac::Hmac;
-#[cfg(feature = "include_simple")]
-use pbkdf2::CheckError;
 use sha1::Sha1;
 
 #[derive(Debug)]
@@ -40,29 +38,4 @@ fn rfc6070() {
         pbkdf2::pbkdf2::<Hmac<Sha1>>(test.password, test.salt, c, &mut buf[..n]);
         assert_eq!(&buf[..n], test.output);
     }
-}
-
-#[test]
-#[cfg(feature = "include_simple")]
-fn test_pbkdf2_simple() {
-    let password = "password";
-
-    let out1 = pbkdf2::pbkdf2_simple(password, 1024).unwrap();
-    let out2 = pbkdf2::pbkdf2_simple(password, 1024).unwrap();
-
-    // This just makes sure that a salt is being applied. It doesn't verify that that salt is
-    // cryptographically strong, however.
-    assert_ne!(out1, out2);
-
-    assert_eq!(pbkdf2::pbkdf2_check(password, &out1[..]), Ok(()));
-    assert_eq!(pbkdf2::pbkdf2_check(password, &out2[..]), Ok(()));
-
-    assert_eq!(
-        pbkdf2::pbkdf2_check("wrong", &out1[..]),
-        Err(CheckError::HashMismatch)
-    );
-    assert_eq!(
-        pbkdf2::pbkdf2_check("wrong", &out2[..]),
-        Err(CheckError::HashMismatch)
-    );
 }
